ResourceFolderRepo init checks file cache dir, and can save
Only saves non-Psi ResourceItems and ResourceFiles (of which
there are currently none). A follow up commit will make
ResourceFolderRepository init use VirtualFile instead of
PsiFile and create non-Psi resource items which can
actually be saved and reloaded.

Loading is automatic from the ctor.

Saving will happen in the next change. We keep track of
file cache freshness stats to determine when to save
a new copy (only when stale), to avoid unnecessarily
updating the file caches.

Bug: b/25078759
Change-Id: I9b7b366e1d6e47368ad55a23d5369aac1ed2629a