Move perf stats and timestamps to GuiTestCase
These were added to GuiTestRunner.runChild by commit 7398d56, and
perhaps should have been in GuiTestCase in the first place, since this
way they can avoid capturing some of the common setup and teardown work.

Anyhow we're aiming to simplify and improve the GUI-test setup by, for
example, moving the rest of GuiTestRunner.runChild to GuiTestCase in
forthcoming changes.

Change-Id: Ia21569239f230b7b99ae200c48485f1d28ad08c0