Remove assumptions on PsiResourceItem, make package private.
Allow plain ResourceItems instead of PsiResourceItems in a few
places. A utility already exists in LocalResourceRepo to
handle either Psi or non-Psi (non-Psi fallback parses
and creates Psi on-demand). With some refactoring, only
the rendering package needs to know about PsiResourceItem.

Also make a few places assume that changes in Psi are
sufficient after a "refreshResources" call. Changes
need to propagate to the underlying file if we are
to re-initialize ResourceFolderRepository from
files instead of Psi, so add a few saveDocument calls
(Psi listener is insufficient because refreshResources
resets everything).

Bug: b/25078759
Change-Id: I74aa866958a21bd391aacd0fe95e1aa373c3bdff