Copying studio skins to sdk and using them
Nexus device skins are included in studio. Previously we used the skin
dir contained in the studio install directly. However, this was
unexpected for users, since studio is often renamed or moved. In this
change the skins are copied from studio into the sdk and then using the
relative path in the SDK.

https://code.google.com/p/android/issues/detail?id=163045

Change-Id: I07e6271c089519d10183d134711fad4cb75b6dda
(cherry picked from commit 3aa5223a692780567200f8430e00e8f625ee0c49)