Show downloadable system images by default.
Show downloadable images by default.
When creating a new AVD and the user need to pick a system image:
if the list is empty because there are no images downloaded or if
non of the downloaded images are recommended, show the images
that can be downloaded.
If there still are no visible system images after getting the
remote images, show the non recommended images as well.

Bugs: http://b.android.com/198662  http://b.android.com/199410

Change-Id: Ia5b0302294ea0d5486447d51b39b79b46712084d