Separate targets from avds, remove MissingTarget.
Previously, an AVD was defined by an AndroidTarget, a tag, a vendor,
and other information. The tag and vendor were used to look up the
SystemImage in the target. The target was never really used for
anything other than that, though, and the necessity to create a
target when the relevant platform or addon was missing caused
needless complexity.
Now the avd contains the system image directly, without using the
target at all.

Fixes https://code.google.com/p/android/issues/detail?id=195141

Change-Id: Ieee7d3855b03956b60cdcd51a2e1407b52a75f8c