Detect scenario where application crashes at startup
After a launch, we monitor the device for the application to come up.
Since we don't know how long that takes, we have a timeout of 10s (right
now). If the application actually started, connected to the IDE, and
then died within that span, we'll still keep waiting. This CL fixes that
scenario.

Change-Id: I7b851064b75c05548df7951fc86cce01022a04a7