Changing "installed size" to "download size" in welcome wizard
Previously we were using a hardcoded or wrong value for the installed
size of welcome wizard components, and we weren't correctly accounting
for which packages would actually be installed. Also, the APIs were
unnecessarily unwieldy.
Changed from "installed size", which we can't really know, to "download
size", which we can. Also cleaned up/simplified the code somewhat.

Mostly fixes
https://code.google.com/p/android/issues/detail?id=175220
Also fixes
https://code.google.com/p/android/issues/detail?id=174902
as a side effect.

Change-Id: I8c75dc8efc7b7c2a57220402122b71727189091e