Use feature level instead of api level when comparing API levels
When passing the device version to gradle, we use the feature level, but
when comparing the data in build info to the device, we use the api
level. This causes a mismatch, which this CL fixes.

Since we are using the feature level, this CL also renames the data
coming out of the build info to be the feature level.

Change-Id: I2ec00d876e47da2d9e3e61ac5805f92cb68ae9bf
(cherry picked from commit 0225d1057ecc36513cc726ce5989577713288eff)