Fixing logging/progress in installer
We were showing a lot of messages during sdk reload in the installer
window. Using a different logger for that part so we don't pick that up.

We were also showing a progress window on top of the normal installer
progress window if you used the sdk manager with no projects open.
Detecting that case and using an EmptyProgressIndicator to prevent that
from happening.

Change-Id: I6f44f3a7c1796a643dc246ed1d9a8f2c01b05674
(cherry picked from commit 65242991e18d526401bce36830c46bf4cf247f7f)