Distinguish whether IR was enabled or not for full APK installs
When collecting deployment stats, all full apk installs were grouped
together without distinguishing whether instant run was enabled or not.
This CL distinguishes between the 2 by using LEGACY for apk installs
where IR was disabled and FULLAPK for when IR is enabled, but we did a
full APK install (such as for a clean build).

In addition, this CL also:
 - removes the build time monitoring as that depends on a number of
   other factors.
 - changes the URL endpoint to use irstats2 (from irstats1)
 - counts the deployment as started only when the launch tasks have
   started, so that failed builds that stop deployment aren't included
   in the count

Change-Id: I07ccf4d601e47ec8c2e76d77b4e0fa7cf7c0b100