Don't run setUp or tearDown after IDE errors
Commit b7ecc29 moved the effective check of doesIdeHaveFatalErrors from
GuiTestRunner.runChild to GuiTestCase.setUp. This simplified the code
but resulted in the @After method being called when the check failed.

Since commit 5d62181 replaced GuiTestCase with GuiTestRule, we can check
doesIdeHaveFatalErrors in the Statement.evaluate method instead, so when
there are IDE errors it produces a AssumptionViolatedException directly
and doesn't call setUp or tearDown.

Change-Id: I3731537023dc3ea89998578b71ae6465790f9b5a