Notify why IR was disabled during a launch
IR can be disabled due to a variety of reasons. This CL exposes the
reason why IR is disabled for a particular launch to the user via a
notification.

This CL also changes all the IR notifications to have a single
"Don't show again" hyperlink. Clicking on that link changes the Instant
Run setting that controls whether notifications are shown.

Fixes https://code.google.com/p/android/issues/detail?id=195466

Change-Id: I29eb4bb862d0ff3153207d9bf0cec6527a2ac4b0