Replace "Host GPU" checkbox with a GPU dropdown with 2 items.
The 2 items in the dropdown should specify whether the graphics are
being rendered by hardware/software and the GLES version.
Also: remove the store snapshot checkbox. We will introduce it
again when it works correctly.

Documentation impact:
AvdManager http://developer.android.com/tools/devices/managing-avds.html
currently does not show the "Verify Configuration" step (same as edit
and AVD). If the newly designed documentation includes this step there
is an update to the controls shown.
The 2 checkboxes "Host CPU" and "Store a snapshot for faster startup"
are gone. There is a dropdown as mentioned above instead.
It is no longer possible to setup a snapshot after running an AVD.

Bug: http://b.android.com/200133

Change-Id: I86e99221b69e8abd9a17ccef42372b702ba4bda0