[UITests] Make GuiTestRuleTest platform independent
On Mac, Intellij's messages/dialogs are rendered via sheets with
underlying JDialogs titled 'This should not be shown'. Unfortunately,
when GuiTestRule closes leftover dialogs, it will not show any
useful information about the dialog beyond that message.

The test is now portable by using plain JDialogs throughout (tested
on Mac, Linux and Win).

I've also documented the implict assumption in GuiTestRule about
the presence of a focus manager to pick the active dialog. Without
a window manager, in Xvfb we only close the first modal dialog.

Change-Id: I93050cdbfbc21c5b4edde155c0461bd9d3cd080a