Differentiate between manifest change and resource change
We see the number of full builds to be approx the same as the # of
hotswaps. The cause seems to be that the manifest or resource referenced
from manifest has changed. This CL differentiates between the two
possibilities in analytics.

(cherry picked from commit 2c01d837156cd78829c45ff913add404bedccf19)

Change-Id: I603656911b0e0b03cf50c591be7a9b2ece2a89a5