Differentiate between manifest change and resource change
We see the number of full builds to be approx the same as the # of
hotswaps. The cause seems to be that the manifest or resource referenced
from manifest has changed. This CL differentiates between the two
possibilities in analytics.

(cherry picked from commit 05847f32b8cbde0b4d2f84f2a2928cd340e4f11f)

Change-Id: I603656911b0e0b03cf50c591be7a9b2ece2a89a5