app launch: remember a setting across studio invocations
When launching an app while the app was previously launched, we show a
dialog that prompts the user whether the existing session should be killed.
This dialog has a "Do not show again" option, but that setting is only saved
for the active Studio session and is lost if you restart Studio.

This CL persists that setting to disk so that the prompt is not shown again.
Fixes https://code.google.com/p/android/issues/detail?id=200710

Change-Id: I4ba3159da82a23ec390b91eba70fb53e7676296f
(cherry picked from commit d559770fa2071eeaf81a5854de4e97b8db535cd9)