Make sure the Configure AVD options can handle a missing device.
BUG: http://b.android.com/200761 (first problem)
There is a possibility that user generated device descriptions
can get corrupt. When this happens all the AVDs that is using
custom devices will end up in a "Repair Device" state.
When the user clicks that link they promptly get a NPE.

This CL makes sure we can repair such an AVD.

Change-Id: I14d18b44ec49b902b9afea029339bb2dd8b2f80b