IR: Always push build timestamps to device
The build id is typically pushed to the device and examined before the
subsequent build to see if it has changed. However, we could get into a
situation where we push an IR enabled apk, then use a variant w/o IR,
and then come back again to the 1st variant. In such a case, the build
ids will match since the non IR enabled variant did not update the
timestamps.

This CL fixes the issue by always pushing the build timestamp, whether
IR is enabled or not. In theory, if you are not using IR at all, then
this is not strictly necessary. It is only a cleanup after IR issue, and
even then, it is only necessary because IR currently leaves state
outside of the apk folder itself.

Fixes https://code.google.com/p/android/issues/detail?id=200807

Change-Id: I626a97ac25ca45c2d7cd7e2c7abc9a57ce726e76