[New PSD] Now we build dependency graph by reading Maven POM files.
The Android Gradle model provides a *flat* list of dependencies,
which is useful for setting up projects but not so much for the new
PSD, given that we need a real graph where transitive relationship
are displayed.

This CL fixes that limitation of the Android Gradle model by
constructing a dependency graph by reading POM files, which are
the descriptors that contain dependency information.

Change-Id: Ieca58c2d650f7d758e9410745dac96fca80f12d9