Fix logic so that "No Filters" not always default on startup
In a previous CL, I refactored code to allow plugins to register
custom logcat filters. In doing so, I made it so the creation of
a "Show only selected application" filter got deferred to later
than usual. This, in turn, caused logic which tried to select the
last known filter by name to fail, thereby leaving "No Filters",
the first item in the list, selected.

Cleaned up the code a bit to separate the "update filter list"
logic from the "select filter" logic.

Change-Id: I9efe0017c205eb0fdf260fe5930c08564f855e6d