Don't force update experimental plugin users
Commit bd340e9ef6b enforced that users are always on the latest version
of the gradle plugin if it was a preview version. This check forces
update of users on the experimental gradle plugin, which is incorrect.

This CL is a quick fix: once a sync is complete and we know the plugin
version, we only do the force update if we know for sure that it is a
non experimental plugin. Unfortunately, the method getPluginGeneration
was not present in 2.0.0-alphaX, so those users won't be force updated.

Change-Id: I09a508d7cdf318af326cda6d5f193ae6260be373
(cherry picked from commit b19f34d92501c2d8c206afdbbf7a34385fa4a189)