Support blacklisting devices with a broken run-as
run-as devices doesn't work on certain Samsung devices, giving an error
for form: "run-as: Package 'foo' is unknown"

In such a case, we cannot support dexswap/freezeswap, i.e. copying of
dex files when the app is not running. This CL adds support for such a
scenario: If the app is not running, and we attempt to run, it will try
to install dex files, and if that fails with the run-as error, it will
blacklist that device for future sessions.

For such black listed devices, we'll perform an "assembleDebug" if the
app is not running. This task will generate an APK with dex files
bundled inside (if there were source or resource changes).

Fixes https://code.google.com/p/android/issues/detail?id=201234

Cherry picked from 772f07c73ce5d3c0f99f1aff9eaa27723a4debd7 and fixed
conflicts in
	android/src/com/android/tools/idea/gradle/run/GradleInvokerOptions.java
	android/testSrc/com/android/tools/idea/gradle/run/GradleInvokerOptionsTest.java

Change-Id: I81c4eb8f47d21f18457a4e772f5acd7d8686a740