Clarify that getResourceNames takes a ResourceType parameter
Previously it was a string parameter, which looked
like it could be a ResourceFolderType as well.
It really should be a ResourceType. Note that the
implementation of getFileResourceNames will convert
the ResourceType to a String and then to a
ResourceFolderType to determine if it's a non-value
resource, but that should be okay.

Namely this refactoring prevents a user from passing
in the string "values", which is a valid folder type
but not ResourceType.

This allows us to remove the fallback/delegation in
LocalResourceManager to super.getResourceNames(),
when the string is not a ResourceType.

Change-Id: Ib15d7023a7ba2b1c104865c45d23cf28e79f2b55