Adding current set of DeveloperServices to the assist panel.
Getting the set of android modules and instantiating DeveloperService
for each. These are passed down to the panel so that it may install
dependencies and check to see if they are installed.

Note that this is an early pass and is insufficient for our needs.
Later cls will listen for file focus in the editor changing and
both announce to the assist panel and provide a new set of
DeveloperService's.

Change-Id: Ib39a4f52bb927c46a3218d6c49f14cb88d0bfd6e