Stop resizing the device preview when there is plenty of space.
When the user resizes the "Select Hardware" dialog the preview on
the right will resize down with the size of the dialog.
The previous code used 10% of dialogonal space as the desired
minimum size of the preview. However this caused the preview to
resize when the user expanded the dialog vertically.

The code has been modified to only compute the size horizontally
which will avoid the resizing of the preview when the dialog is
expanded vertically. This works because the preview pane is kept
at a constant width.

Change-Id: I7c7e9b282f8713018026449001c3572b226db76a