Don't show the cause for clean build (build timestamps don't match)
It turns out that we do a clean build only for 2 cases:
 1. Build timestamps don't match.
 2. The API level of the build on disk doesn't match the device.

The second scenario was a safeguard necessary in old implementations
when launching to an AVD: during the initial launch, we did not know the
API level of the emulator, so the build was targeting API 1. But the
subsequent launch to the same emulator would return a different API
level (but same build timestamp). This workaround is not necessary
anymore since we know the API level of the AVD even before it is
running.

The first scenario cannot happen during an instant run: if the
deployment is to the same device, then the only way for the timestamps
to differ is if the user manually uninstalled the app on the device.

But the current message (Build timestamps don't match) was being
displayed on all initial runs, because very likely they don't match.
That did not make any sense since (the expected UX is that these
notifications should only be shown for the instant run, not for the
first run).

This CL avoids this issue by completely eliminating this notification
(cause for a clean build) as it is not useful in both the scenarios.

Fixes https://code.google.com/p/android/issues/detail?id=201782

Change-Id: I4936c9fa864b51e0115f7fea785c0466a8ff5e2f
(cherry picked from commit 5a41148e9a8e44b8c96ebd5d48d00c42011598f6)