Use the same manifest files to decide whether a full build is needed
When deciding whether to do a full build, we compare:
 - the timestamps of all the source manifests to
 - the timestamp of the what was installed (merged manifest)

However, it is simple enough for the source timestamps to change (any
action followed by an undo), but in such a scenario, the merged
manifest's timestamp will not be updated. Once this happens, we never
recover until there is a real manifest change causing us to
unnecessarily build all the time.

This CL changes the comparisons to always be on the source manifest.
When an APK is installed, we cache the last modification time of the
source manifests. Since the comparisons are now all on the source
manifest, any changes to the source will be caught and cached, and there
is no dependency on what the manifest merger does.

Fixes https://code.google.com/p/android/issues/detail?id=201880

Change-Id: Ib40308361781d6ee5fc506371b2675d334f8a469
(cherry picked from commit 8560e63b5990cd142ebd9ced61a6aa80908cb516)