Don't terminate the process handler when running tests
When running tests, there are two events going on in parallel that leads
to a race:
 - The process handler waits for the test process to die, and when that
   happens, it terminates the process handler, which effectively results
   in the test session being terminated.
 - AndroidTestListener meanwhile keeps monitoring the output of the "am
   instrument" command to detect the test pass/failure and completion.
   When the test suite completes, it informs the UI that the tests are
   done.

If the first scenario happens before all the test results are
received/processed by the test listener, the UI ends up showing that the
tests were terminated instead of saying that the tests completed.

This CL fixes this issue by making the process handler not monitor the
state of the remote process for test runners.

Fixes https://code.google.com/p/android/issues/detail?id=201968

Change-Id: Ic0a7f2a71d8886403efe8862893fd893bdaacab9