Use runReadActionInSmartMode() to avoid IndexNotReadyException
When launching an application, we need to determine the default activity
that is used. This requires indices, and may fail if indexing is in
progress.

This CL replaces the read action with a runReadActionInSmartMode to
avoid the exception if indexing was in progress.

Fixes https://code.google.com/p/android/issues/detail?id=200427

Change-Id: Ic9cc1ba1cf3afdd910316fca4b2c2bac548edb8f
(cherry picked from commit c6e23be08acaec3eef2b53df8d1da5f471c32cc1)