[New PSD] Now 'Android Test' and 'Test' artifacts show dependency on the 'Main' artifact.
This is displayed in the 'Resolved Variants' view.

Both 'Android Test' and 'Test' depend on the 'Main' artifact in the
same variant. This means that dependencies in the 'Main' artifact
are inherited by the test artifacts.

This CL makes that dependency explicit.

Also, when the 'main' artifact is selected, all other instances of
the 'main' artifact, in the same variant, as automatically selected
as well.

Change-Id: Ic55bc48d3b560a3effa92b3c7116dc576b2bbbee