[New PSD] Now dependency scopes are properly mapped to build artifacts.
The new PSD has a bug where parsed dependencies are matched against
dependencies coming from the Gradle model only by path (in the case
of modules) and group:name (in the case of libraries) but the scope
of the dependency was never used. This bug caused dependencies to
show up twice in the "Declared Dependencies" section, if, for
example, the same dependency was declared in 2 scopes:
"androidTestCompile" and "testCompile".

To fix this bug, the new PSD used the scope of the dependency to
determine the build artifact the dependency belongs to, as follows:
- if the scope ends with "androidTestCompile" or
  "AndroidTestCompile", the dependency belongs to the "Android Test"
  artifact
- if the scope ends with "testCompile" or "TestCompile", the
  dependency belongs to the "Unit Test" artifact
- if the scope ends with "compile" or "Compile", the dependency
  belongs to the "main" artifact

Change-Id: I3791a2e69c4c5df73568d3c15a60c39503f98e2c