[New PSD] Implemented "Group" dependencies by Variant and Artifact.
Dependencies are grouped as follows:
1. 2 or more variants have the same dependencies for a specific
   artifact type ("main", "unit test", "android test")
2. when grouping non-main artifacts, the dependencies in the
   main artifacts, in the same variant, are also checked

Example 1:
The following structure

 - variant1_main
   - library1
   - library2
 - variant2_main
   - library1
   - library2
 - variant3_main
   - library3

will create these groups:

 - variant1_main, variant2_main
   - library1
   - library2
 - variant3_main
   - library3

'variant1_main' and 'variant2_main' grouped because they have
the same dependencies.

Example 2:
The following structure

 - variant1_main
   library1
   library2
 - variant1_test
   library3
 - variant2_main
   library1
 - variant2_test
   library3

In this case 'variant1_test' and 'variant2_test' are not grouped,
even though they both declare the same dependency ('library3'). The
reason is that each test artifact depend on the 'main' artifact,
and in this case 'variant1_main' and 'variant2_main' do not have
the same dependencies.

Change-Id: I384ee080282991975002c52ea0ff4c85439639e8