[New PSD] Now the PSD shows dependencies with versions higher than the declared ones.
This is the case when the build.gradle file declares a version of
a dependency, but Gradle resolves it to a newer version to resolve
a version conflict.

See: https://docs.gradle.org/current/userguide/dependency_management.html

Example:
Module 'app' depends on module 'lib'. Module 'app' declares that
it depends on guava 16, and module 'lib' declares that it depends
on guava 19. To resolve that version conflict, Gradle will make
'app' depend on guava 19.

The new PSD uses the same notation as Gradle to indicate that a
dependency version has been ugpraded, in this case:

  guava:16->19

(See screenshot)

Change-Id: I7ddb065da43781d45c4533b1e0aa3de6171eb24b