ModelWizardDialog can handle ModelWizard being cancelled externally
ModelWizard has a public "cancel" API method, but currently if a
ModelWizardDialog wrapped a ModelWizard but then someone else called
cancel, the dialog wouldn't know, and next time the user would take
an action, the wizard would throw an invalid state exception.

Now, ModelWizard reports, via a listener, when it is finished or
cancelled, and ModelWizardDialog hooks into that API to respond
accordingly.

The Wizard.Facade API was also updated to allow child steps to
cancel the wizard, if necessary.

Change-Id: I9209c12206702998cce0b2eda93efb1ad251ed6e