AvdDeviceData: Add Width/Height/diagonal value constraints
In the AVD manager, while changing the values of the
Width/Height/Diagonal, its possible for the model data to become
"corrupted" with invalid values. This was shown by an UI test that
was failing with and assertion inside
AvdDeviceData@getDeviceScreenDimension(), where width/height
were < 0.
To avoid this problem, we now added a constraint to this properties,
that will force them to always have valid values.

Change-Id: I5e3ab609013c82a231522e5b5ad06127cd86ba89