Whitelist and allow hotswap for certain multi-process apps
Using multiple processes in an app disables hotswap and forces cold
swap. This CL tweaks that logic and allows hotswaps in the common
scenario where an app is forced to use a 2nd process due to the presence
of a 3rd party library such as leakcanary. In such a case, we don't care
that much about updating the 2nd leakcanary process.

Change-Id: I9dbcaa221b5b941007a4ae518c9b6faaf85001c0