Add layoutlib in Android Studio
This CL doesn't change anything in the API or the Bridge and it can
still work with the layoutlib.jar versions from the SDK. It can be
disabled at runtime setting the "use.sdk.layoutlib" property to true.

Some tests will now use the internal resources distributed with the SDK
so they needed to be updated.

Change-Id: I05c328198c209c61d06c1e806bdf86bb0346720f