Change installers to install listeners where possible
Previously, when we wanted some special actions to happen during sdk
component installation, we created a new subclass of PackageInstaller.
However, it turns out that we need more complex composition of behaviors
than this allowed: for example, we want to be able to apply diff-based
updates to maven packages, etc.

Change-Id: Iaa33d86177a82dad1148b563eca6eda687a17d62