Fixing an issue with source generation after project sync.
The max module count preference need to be taken into consideration
only when the skip source generation preference is enabled. But,
right now it's taken effect even when that prefercne is not enabled.

This chnages fixes that issue, so that source generation is enabled
by default and user can diable it with the preference.

Change-Id: I85385f2469bb2bcccde2f158f2331de5eee3787c