UI Test Framework: Keep running with timeouts on tearDown
TearDown calls waitForBackgroundTasks() =>that can throw a timeout,
for example when there is a background task that is blocked on a
dialog ("Are you sure you want to close the project while gradle is synchying?").
If there is a timeout, then cleanUpAndCheckForModalDialogs() and
closeAllProjects() don't run and none of the next tests are allowed to run.

Jenkins eventually times out after 180min.

Change-Id: Ic7e2ad7423df6bd590c72d1c8aa7211284430ef0