Worked around incompatibility between builder-project and cached model.
When updating Android Studio to use builder-model 2.2.0-dev there
is an incompatibility between the 'Library' interface and the
cached model, resulting in a UnsupportedMethodException.

Gradle sync is smart enough now that it can detect this and
trigger a full sync. The problem is that AppResourceRepository
attemps to use the cached model before sync happens.

This CL works around this timing issue, by simply catching the
UnsupportedMethodException if it happens before sync. Once sync
is finished, the state of the IDE will be fully restored.

Change-Id: I493caa24a27aa76da5523c721f5da287037ed1cd