Tweak ResourceFolderRepository cache saving/updating heuristic
Make it happen after fewer files are modified.

Previously it tolerated reparsing 75% of the files
before it updated the cache (had to be really stale).
Instead flip it to 25%. Previous measurements (on
linux) showed that loading from cache is ~3x faster
than without the cache.

Also change the accounting to not count files that are
explicitly skipped for whatever reason (data binding,
or density-based resources), since there's no chance
those will be cached anyway.

Bug: b/25078759
Change-Id: I862d2525ce15257995c233009b1667e7381dc4cd