Now we let users manually update Android plugin version.
There are cases where the automatic update fails (e.g. when
build files use Groovy scripting to set the plugin version)
and closing the project is too aggressive.

This CL fixes that by giving the user the option to
manually update her build files to use the latest plugin.

If the user chooses to update manually, we force a
sync failure.

Change-Id: Ib4e6066ec5ed9452ee31d8dc1bd19adf56768861