hv: Don't show window picker if there is only 1 window
When hv is invoked, we show a dialog to select the window whose view
hierarchy is being inspected. This is not needed if there is only 1
window. This CL cleans this up by first obtaining the list of windows,
and showing the WindowPicker only if there are multipl windows.

As a side effect, we don't need a progress bar in the window picker
since it now knows the list of windows to display in its combo box
when it is constructed.

Change-Id: Id2394975c0e0fca68e05e897b33039cfd79629a5