Fix already installed services triggering reinstalls
Normally, when a service in Project Structure is unchecked and you
check it, that puts the service into "edit" mode. If you uncheck it,
it restores it to unmodified. If you hit "OK" on the dialog when a
service is in "edit" mode, it installs it.

However, when you open the project structure dialog with services
that are already installed, they are initialized to checked,
programatically.

Unfortunately, some logic in the developer services UI was failing to
distinguish between the different reasons the checkbox might have
been checked, particularly: "user wants to edit this" vs.
"the service was already installed"

Unfortunately, what's happening now is that, each time the user opens
the project structure dialog with a service installed, the code
detects it is checked and thinks the service entry as "modified".
When the user hits OK to close the dialog, it tries to reinstall the
service.

This is mostly harmless, except that it messes up our usage tracking
statistics and causes the project to reindex.

Change-Id: I360f43e224e40edee214178521fb38d020711e10