NewProjectTest: testNoWarningsInNewProjects flaky
This test, and others, wait for the gradle sync action button to
display and be enabled... But a new button is added and removed over
time, and we may find a "dead" button, that will never be enabled.
Now we keep looking for the button until all the conditions are meet.

Change-Id: If91f8371858ba65221eedefd6e922bdcf7050d01