Track instant run timings by deploy type
This CL keeps track of the timing of instant run deployments (of all
deployment types - hotswap, fullapk, dex, ..). These stats are uploaded
when the session changes. A typical URL looks like:
  HOTSWAP=1a1a1&FULLAPK=19&timeUnit=seconds
Note: There is an "a" instead of a comma to avoid URL escaping turning a
single char into three.

Change-Id: I36bb620c46aec6f2dec7a4755f695e110c064eba