Fix Rerun action to use the status of currently selected configuration
The Rerun action was enabled or disabled by looking at the status of the
current descriptor (maps to the console) instead of the currently
selected run configuration. As a result, if you had the console of some
previous debug session currently active, then Rerun was being disabled.

This CL fixes it to look at the selected run config instead of the
descriptor.

Fixes https://code.google.com/p/android/issues/detail?id=204775

Change-Id: I66c1bc5322929d113ebccd2063dbd78d99857c50