Fix TestManifestTest
The test was broken because it was opening the manifest file with
the default tab and it was assuming the default tab to be Text.
But if the test runs after another test that finished on another
tab (namely ManifestEditorTest that is changing the tab to Merged
Manifest) then that setting is saved and becomes the default for
the next time that file is opened.

Change-Id: I21f1b72730378caa55df972caf93fe6f68fd0578