[New PSD] Now selecting multiple dependencies clears the "Target Modules" view.
This CL fixes a bug where selecting different dependencies in the
"All Dependencies" view (e.g. Guava and JUnit) will result in the
"Target Modules" view to show the target modules of the first
selected module, which is wrong, because that information does
not apply to the rest of selected dependencies.

Change-Id: I08ed86ad86cf352bfa051b892607d27bf7787ffc