a11y: UpdaterTreeNodes behave like regular checkboxes
Given UpdaterTreeNode instances are rendered as cells inside a table,
screen readers don't seem to be able to "dive" into the JPanel
accessibility children to find the checkbox and label. The fix
is to create a custom AccessibleContext for the JPanel (the Renderer)
that exposes a mix of the underlying checkbox and label so that
it behaves like a regular checkbox for screen readers.

Change-Id: I7b145d6cc5ff1b795525d40d95b9e839d4109d34