Use the same enum editor in both forms of property sheet
The custom inspector panel for layout width and height were just simple
text boxes. The table based property sheet provides a combobox with a
dimension resource selector.

This CL refactors the existing NlEnumEditor into a core component that
doesn't depend on being embedded as a cell editor, and uses the same
component inside the custom inspector.

Change-Id: I7bf57354d109dcee8463552a16e75d9108e1a819