Remove calls to focus in UI tests when not needed
A lot of calls to focus were simply giving enough time for
the ok button to become enabled. So simply wait for that if
that is the case.
The other ones were not needed, and should be removed because
they could potentially hide bugs like wrong preferred focus
component...
Overall, calls to focus should be limited to times when it
actually makes sense.

Change-Id: Iee87aba8e3ed4dd0b41bc84d1595bce250915a31