Remove AndroidFacet reference from AssetRepository on dispose
When a project is disposed, the AssetRepository keeps a reference to the
AndroidFacet of the disposed project. LayoutLib is not unloaded until
studio is closed so this reference is kept around.
This is usually not a problem since the new project will overwrite that
reference but it generates an error in the unit tests.
This CL just makes sure that we remove the reference so unit tests are
happy.

Change-Id: I27581162a6b2d7ebf6db12502fd14329c7ebe51e