Move calls to getPrimaryResourceDir closer to Dialogs
Moving calls closer to Dialog code moves us closer to
having UI that disambiguates where to create new
resources. Otherwise, if a module has multiple resource
directories the "primary" choice may be unexpected.

This focuses on moving calls within AndroidResourceUtil
out. It also attempts to focus the "output" of a
Dialog to being the resource directory instead of the
module or source provider, when possible (usually what
the module or source provider is for). This prevents
the outside code from trying to interpret the resource
directory differently from what the dialog intended.

May eventually help:
Bug: b/27775947

Change-Id: Id2cebdd1d353838b0a82562f2cb6bf0ad3114088