Fix ResourceFolderRepositoryTest#testSerializationAddXmlFile
...AndLoad. The assertion was wrong (copy-paste bug),
asserting on the original "resources" and not the
"fromBlob". That used to work before the IJ 2016 merge
by luck, since the PSI listener is active and fills
in the "resources".

Move the reset() call earlier so that the listener
isn't active and assert on "fromBlob" instead.

Add a workaround for copyFileToProject in another test
case testDeleteResourceFile. copyFileToProject has
changed to first create an empty file, then copy the
data over. This trips up the PSI listener.

Change-Id: If500f999ecfef4948c0d4269426f1ffd98ee267e