ThemeEditorTableTest: testResourceCompletion
Throwing disposed exception when running the second time.

The old editor file is found on the list of recent closed files, and the
exception is thrown when ThemeEditorVirtualFile.getPath() is called and
a call is done project.getBasePath(), as the project was already disposed.

This class was also memory leaking the project, so the hard reference
was removed all together.

Change-Id: I52b46c4a71a5d5885ea3eceb168c4ccc7eb76a3a