Simplify the inheritance for create res dir and file dialogs
Make CreateResourceDirectorDialog a real class instead
of abstract by passing in a function to customize
the desired validator instead of using inheritance.

Make CreateResourceFileDialog the thing used instead
of a "MyDialog" inner class. Also pass in a validator
factory, and have the validator constructor set
the myRootElement instead of overriding doOKAction.

Simplify the dialog interface to just have
dialog#getCreatedElements instead of a bunch of
other getters.

This will give us flexibility to choose the impl of
the dialog instead of always creating the specific
dialogs.

BUG: b/27775947
Change-Id: Icb0c871be09dc4d9a9a6d18068f1b8fb91da2a75