[New PSD] Fixed mapping of configuration names to build artifacts.
This CL fixes a bug where the new PSD mapped build artifacts to the
wrong configuration name. For example, for the build artifact
"Unit Test" in the variant "f1f2Debug" (where "f1" and "f2" are
flavors and "debug" is the build type), the new PSD was mapping this
artifact to the configuration name "f1f2DebugTestCompile". This
bug resulted in dependencies not being displayed in the new PSD.

With this CL, PsAndroidArtifact (the model that represents a
build artifacts) correctly calculates all *possible* configuration
names. Using our previous example, the configuration names are:
  - testCompile
  - testDebugCompile
  - testF1Compile
  - testF2Compile

Change-Id: Ia8f2d2fc43a900d8edeb4dae79a540efaf269c67