Reduce painting and paint constraints on both surfaces
Moves most of the functionality from the BlueprintLayer to a
ConstraintLayer that can be overlaid on top of both the blueprint and
the regular mode.
Also, added clipping detection to avoid painting calls on the regular
surface while animating the blueprint one.

The constraint layer will also invalidate the surface that is actually
being painted on.

Change-Id: Id3abdf6b3d620c1d2e57a3d50545ada4bf63a3ed