[New PSD] New implementation of "Android Scopes" panel.
This is the panel where users select the scope of a
new dependency to add. To make it easy to the user,
this panel shows all possible scopes (compile, testCompile,
androidTestCompile) and the list of build types and
product flavors.

The user only has build theh scope(s) by selecting the
elements in this panel, removing any guesswork required
when she has to type the scope directly into build.gradle
files.

This panel is not fully functional yet, but the UI is
pretty much final.

Change-Id: I70d12183f4cea986ae5ac5e0eb91fef26aa6423b